authorization for buttons and routes

+++ routes/campgrounds.js
@@ -1,21 +1,9 @@
 const express = require('express');
 const router =  express.Router();
 const Campground = require('../models/campground');
-const ExpressError = require('../utils/ExpressError')
 const catchAsync = require('../utils/catchAsync')
 const { campgroundSchema }  = require('../schemas')
-const { isLoggedIn } = require('../middleware');
-
-// Middleware for serverside validation
-const validateCampground = (req,res,next) => {
-    const {error} = campgroundSchema.validate(req.body);
-    if(error){
-        const msg = error.details.map(el => el.message).join(',')
-        throw new ExpressError(msg, 400)
-    }else{
-        next();
-    }
-} 
+const { isLoggedIn, isAuthor, validateCampground } = require('../middleware');
 
 router.get('/' ,catchAsync( async (req, res)=>{
     const campgrounds= await Campground.find({});
@@ -27,16 +15,23 @@ router.get('/new' , isLoggedIn , (req,res) =>{
     res.render('campgrounds/new' )
 })
 
-router.post('/' , validateCampground, catchAsync(async(req,res,next) =>{
+router.post('/' ,isLoggedIn, validateCampground, catchAsync(async(req,res,next) =>{
     // if(!req.body.campground) throw new ExpressError('Invalid Campground Data' , 400)
     const campground = new Campground(req.body.campground);
+    campground.author= req.user._id;
     await campground.save();
     req.flash('success' , 'Succesfully make a new campground!')
     res.redirect(`/campgrounds/${campground._id}`);
 }))
 
 router.get('/:id' , catchAsync(async (req,res)=>{
-    const campground = await Campground.findById(req.params.id).populate('reviews');
+    const campground = await Campground.findById(req.params.id).populate({
+        path:'reviews',
+        populate: {
+            path: 'author'
+        }
+    }).populate('author');
+    console.log(campground);
     if(!campground){
         req.flash('error' , 'Cannot find that campground!')
         return res.redirect('/campgrounds');
@@ -44,8 +39,9 @@ router.get('/:id' , catchAsync(async (req,res)=>{
     res.render('campgrounds/show' , { campground });
 }))
 
-router.get('/:id/edit' ,catchAsync( async(req,res) =>{
-    const campground = await Campground.findById(req.params.id)
+router.get('/:id/edit' ,isLoggedIn , isAuthor, catchAsync( async(req,res) =>{
+    const { id } = req.params ;
+    const campground = await Campground.findById(id)
     if(!campground){
         req.flash('error' , 'Cannot find that campground!')
         return res.redirect('/campgrounds');
@@ -53,15 +49,20 @@ router.get('/:id/edit' ,catchAsync( async(req,res) =>{
     res.render('campgrounds/edit' , { campground });
 }))
 
-router.put('/:id' , isLoggedIn, validateCampground, catchAsync( async(req,res) =>{
+router.put('/:id' , isLoggedIn, isAuthor, validateCampground, catchAsync( async(req,res) =>{
     const {id}  = req.params;
     const campground = await Campground.findByIdAndUpdate(id, {...req.body.campground} );
     req.flash('success','Successfully update Campground!')
     res.redirect(`/campgrounds/${campground._id}`);
 }))
 
-router.delete('/:id' ,catchAsync( async(req,res) =>{
+router.delete('/:id' , isLoggedIn, isAuthor,catchAsync( async(req,res) =>{
     const {id} = req.params;
+    const campground = await campground.findById(id)
+    if(!campground.author.equals(req.user._id)){
+        req.flash('error', 'You do not have permission')
+        return req.redirect(`/camgrounds/${id}`)
+    }
     await Campground.findByIdAndDelete(id);
     req.flash('success' , 'Successfully deleted campground')
     res.redirect('/campgrounds')
